+++ main.py
@@ -4,15 +4,14 @@ import securescaffold
 
 from flask import render_template, send_from_directory, session
 from authlib.integrations.flask_client import OAuth
-from google.cloud import secretmanager
-from auth import create_auth_blueprint, register_google_oauth
+
+from app.auth import create_auth_blueprint, register_google_oauth
 
 # flask app setup
 app = securescaffold.create_app(__name__)
 
 # init oauth
 oauth = OAuth(app)
-
 google_oauth = register_google_oauth(oauth)
 
 # create auth blueprint from factory
